# Better performance for reading dimensions of SVG images