With this PR `Image::getDimensions()` reads SVG files only partially to save IO, parsing time and memory. And it doesn’t call `getimagesize()` on SVG images anymore.

### Todo:

* [x] Check if the wrapper [`compress.zlib://`](http://php.net/manual/en/wrappers.compression.php) is supported on all platforms
* [x] Write tests