@ausi 2016-11-16
A quick performance test looks promising:

| Size (Bytes) | images/sec before | images/sec now |
|-------------:|------------------:|---------------:|
|          100 |              7699 |           7518 |
|        1,000 |              7345 |           7532 |
|       10,000 |              5841 |           7557 |
|      100,000 |              1888 |           7573 |
|    1,000,000 |               216 |           7583 |