@leofeyer 2016-11-17
> Check if the wrapper [`compress.zlib://`](http://php.net/manual/en/wrappers.compression.php) is supported on all platforms

I think it is: http://php.net/manual/de/ref.zlib.php#37482