@ausi 2016-11-17
If we want to be sure we could add a `in_array('compress.zlib', stream_get_wrappers())` check.