@leofeyer 2016-11-17
Maybe in the Contao check or Contao manager but rather not in the core. On a page with 100 images, the check would be executed 100 times otherwise.